refactor: 패키지 분리
OutputView 추가 - View 로 묶기
Application -> LadderController, Controller 패키지에 넣음
Ladder -> Model 패키지에 넣음
