+++ src/main/java/kr/codesquad/Main.java
@@ -1,7 +1,9 @@
 package kr.codesquad;
 
+import kr.codesquad.controller.LadderController;
+
 public class Main {
     public static void main(String[] args) {
-        new Application().run();
+        new LadderController().run();
     }
 }
